# 开发技术文档/02-API接口定义文档/openapi.yaml
openapi: "3.0.3"
info:
  title: 知微 ziwi SaaS 平台 API
  description: |
    知微(SaaS)平台 — Phase 1 核心业务 API 文档

    ## 认证方式
    - **Bearer JWT**: `Authorization: Bearer <token>` — 用户身份认证
    - **X-Tenant-Id**: 请求头 — 多租户识别（SaaS 模式必传）
    - **X-Deploy-Mode**: 请求头 — 部署模式: saas / onprem / multi-factory
    - **X-Factory-Id**: 请求头 — 工厂ID（multi-factory 模式必传）
    - **X-License-Key**: 请求头 — 许可证密钥（onprem 模式必传）

    ## API 版本
    - URL 路径版本: `/api/v1/{resource}`
    - 同一接口保留 2 个版本，v1 废弃前 3 个月发送 `X-API-Deprecated` 头

    ## 分页规范
    - `page`: 从 1 开始
    - `page_size`: 默认 20，最大 100
    - `total`: 总记录数
    - 响应格式: `{ "items": [], "total": 0, "page": 1, "page_size": 20 }`

    ## 错误响应格式
    ```json
    {
      "code": "400-1001",
      "message": "参数校验失败：必填字段缺失",
      "request_id": "req_abc123"
    }
    ```

    ## 错误码清单
    | code | 说明 |
    |------|------|
    | 400-0000 | 请求参数错误 |
    | 400-1001 | 参数校验失败：必填字段缺失 |
    | 400-1002 | 参数校验失败：格式错误 |
    | 400-1003 | 参数校验失败：值超出允许范围 |
    | 400-2001 | 业务规则冲突：库存不足 |
    | 400-2002 | 业务规则冲突：状态不允许此操作 |
    | 400-2003 | 业务规则冲突：数据已存在（重复） |
    | 401-0000 | 未认证 |
    | 401-1001 | Token 无效或已过期 |
    | 401-1002 | Token 格式错误 |
    | 401-1003 | OAuth 授权失败 |
    | 403-0000 | 无权限 |
    | 403-1001 | 模块未购买（租户未启用该模块） |
    | 403-1002 | 用户角色无此操作权限 |
    | 403-1003 | 数据范围权限不足（不可跨工厂操作） |
    | 404-0000 | 资源不存在 |
    | 404-1001 | 查询的资源 ID 不存在 |
    | 404-1002 | 关联的父资源不存在 |
    | 409-0000 | 数据冲突 |
    | 409-1001 | 并发修改冲突（乐观锁版本号不一致） |
...
    | 429-0000 | 请求过频繁 |
...
servers:
...
paths:
...
  /api/v1/users:
    get:
      tags: ["M00-用户管理"]
      summary: 用户列表
      operationId: listUsers
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 用户列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserInfo"
                  total:
                    type: integer
    post:
      tags: ["M00-用户管理"]
      summary: 创建用户
      operationId: createUser
      security:
        - BearerAuth: []
        - TenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: 创建成功
  /api/v1/users/{user_id}:
    get:
      tags: ["M00-用户管理"]
      summary: 用户详情
      operationId: getUser
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 用户详情
    put:
      tags: ["M00-用户管理"]
      summary: 更新用户
      operationId: updateUser
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 更新成功
    delete:
      tags: ["M00-用户管理"]
      summary: 删除用户(禁用)
      operationId: deleteUser
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 删除成功
...
  /api/v1/equipment:
    get:
      tags: ["M02-设备"]
      summary: 设备列表
      operationId: listEquipment
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: category_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: factory_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 设备列表
    post:
      tags: ["M02-设备"]
      summary: 创建设备
      operationId: createEquipment
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/equipment/{equipment_id}:
    get:
      tags: ["M02-设备"]
      summary: 设备详情
      operationId: getEquipment
      security:
        - BearerAuth: []
        - TenantId: []
    put:
      tags: ["M02-设备"]
      summary: 更新设备
      operationId: updateEquipment
      security:
        - BearerAuth: []
        - TenantId: []
    delete:
      tags: ["M02-设备"]
      summary: 删除设备
      operationId: deleteEquipment
      security:
        - BearerAuth: []
        - TenantId: []
...
  /api/v1/maintenance-plans:
    get:
      tags: ["M02-保养计划"]
      summary: 保养计划列表
      operationId: listMaintenancePlans
      security:
        - BearerAuth: []
        - TenantId: []
    post:
      tags: ["M02-保养计划"]
      summary: 创建保养计划
      operationId: createMaintenancePlan
      security:
        - BearerAuth: []
        - TenantId: []
...
  /api/v1/maintenance-tasks:
    get:
      tags: ["M02-保养任务"]
      summary: 保养任务列表
      operationId: listMaintenanceTasks
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: equipment_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: task_type
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 任务列表
    post:
      tags: ["M02-保养任务"]
      summary: 创建保养任务
      operationId: createMaintenanceTask
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/maintenance-tasks/{task_id}:
    get:
      tags: ["M02-保养任务"]
      summary: 任务详情
      operationId: getMaintenanceTask
      security:
        - BearerAuth: []
        - TenantId: []
    put:
      tags: ["M02-保养任务"]
      summary: 更新任务
      operationId: updateMaintenanceTask
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/maintenance-tasks/{task_id}/status:
    put:
      tags: ["M02-保养任务"]
      summary: 变更任务状态
      operationId: changeMaintenanceTaskStatus
      security:
        - BearerAuth: []
        - TenantId: []

  # ==========================================
  # M02 — 备件
  # ==========================================
  /api/v1/spare-parts:
    get:
      tags: ["M02-备件"]
      summary: 备件列表
      operationId: listSpareParts
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: keyword
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 备件列表
    post:
      tags: ["M02-备件"]
      summary: 创建备件
      operationId: createSparePart
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/spare-parts/{part_id}:
    get:
      tags: ["M02-备件"]
      summary: 备件详情
      operationId: getSparePart
      security:
        - BearerAuth: []
        - TenantId: []
    put:
      tags: ["M02-备件"]
      summary: 更新备件
      operationId: updateSparePart
      security:
        - BearerAuth: []
        - TenantId: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token，从登录接口获取"
    TenantId:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: "租户ID（SaaS模式必传）"
    DeployMode:
      type: apiKey
      in: header
      name: X-Deploy-Mode
      description: "部署模式: saas / onprem / multi-factory"
    FactoryId:
      type: apiKey
      in: header
      name: X-Factory-Id
      description: "工厂ID（multi-factory模式必传）"
    LicenseKey:
      type: apiKey
      in: header
      name: X-License-Key
      description: "许可证密钥（onprem模式必传）"
  schemas:
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        real_name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar_url:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        real_name:
          type: string
        email:
          type: string
        phone:
          type: string
        role_ids:
          type: array
          items:
            type: integer
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      description: 每页条数，默认 20，最大 100
      schema:
        type: integer
        default: 20
        maximum: 100
        minimum: 1
...